add cors

+++ controllers/productController.js
@@ -22,25 +22,29 @@ export const saveProduct = async (req,res) => {
     const product = new Product(req.body);
     try {
         const savedProduct = await product.save();
-        res.status(200).json(savedProduct);
+        res.status(200).json({message:"Data Sukses tersimpan"});
     } catch (error) {
         res.status(400).json({message: error.message});
     }
 }
 
 export const updateProduct = async (req,res) => {
+    const cekId = await Product.findById(req.params.id);
+    if(!cekId) return res.status(404).json({message:"Data tidak ditemukan"})
     try {
         const updatedProduct = await Product.updateOne({_id:req.params.id}, {$set:req.body});
-        res.status(200).json(updatedProduct);
+        res.status(200).json({message:"Data Sukses terupdate"});
     } catch (error) {
         res.status(400).json({message: error.message});
     }
 }
 
 export const deleteProduct = async (req,res) => {
+    const cekId = await Product.findById(req.params.id);
+    if(!cekId) return res.status(404).json({message:"Data tidak ditemukan"})
     try {
         const deletedProduct = await Product.deleteOne({_id:req.params.id}, {$set:req.body});
-        res.status(200).json(deleteProduct);
+        res.status(200).json({message:"Data Sukses terhapus",});
     } catch (error) {
         res.status(400).json({message: error.message});
     }
